+++ example/lib/native_app.dart
@@ -11,6 +11,7 @@ import 'pages/checkbox_page.dart';
 import 'pages/chip_page.dart';
 import 'pages/color_page.dart';
 import 'pages/group_page.dart';
+import 'pages/header_box_page.dart';
 import 'pages/header_page.dart';
 import 'pages/icon_page.dart';
 import 'pages/input_trigger_page.dart';
@@ -218,6 +219,10 @@ class MyApp extends StatelessWidget {
                             _DemoEntry(
                               'Header',
                               HeaderPage(),
+                            ),
+                            _DemoEntry(
+                              'Headerbox',
+                              HeaderBoxPage(),
                             ),
                             _DemoEntry(
                               'Modal',