+++ lib/src/theme/styles/sbb_styles.dart
@@ -20,3 +20,4 @@ export 'src/sbb_slider_style.dart';
 export 'src/sbb_switch_style.dart';
 export 'src/sbb_text_field_style.dart';
 export 'src/sbb_text_style.dart';
+export 'src/sbb_headerbox_style.dart';